// apps/web/src/lib/api/generated/model/leadOutStage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type LeadOutStage = typeof LeadOutStage[keyof typeof LeadOutStage];


export const LeadOutStage = {
  new: 'new',
  contacted: 'contacted',
  quoted: 'quoted',
  won: 'won',
  lost: 'lost',
} as const;